refactor(lowy): split CodeFilterBar into ModeChipPicker + FileSearchInput
Lowy review on this PR (issue #794) flagged that CodeFilterBar
conflated two unrelated volatility axes:

* Mode selection — sequence volatility, changes when a mode is
  added or removed.
* Filename filter input — activity volatility, changes when search
  ownership, algorithm, or placement shifts.

They lived in one component because they share chrome space, but
the seam was visible in this PR itself: the search-ownership change
touched CodeFilterBar even though mode-selection logic didn't.

Splits the filter chrome into two focused presenters:

* `ModeChipPicker.tsx` — chip + popover, parameterized by `modes`
* `FileSearchInput.tsx` — input + clear button, parameterized by
  `value` + `onChange`

`CodeTab` now arranges them inline in a one-line toolbar div, so
visual co-location is preserved. The `ModeOption` type re-exports
from ModeChipPicker.

Closes #794.
